Time: 3 ms (41.07%) | Memory: 7.7 MB (8.34%) - LeetSync

+++ 476-number-complement/number-complement.cpp
@@ -1,11 +1,11 @@
 class Solution {
 public:
     int findComplement(int num) {
-        long long k=num;
-        long long ans=0;
+        int k=num;
+        int ans=0;
         string s="";
         while(k){
-            long long d=k&1;
+            int d=k&1;
             if(d==1){
                 s='0'+s;
             }
@@ -15,15 +15,15 @@ public:
             k>>=1;
         }
         cout<<s<<' ';
-        long long n=s.size();
-        long long b=0;
+        int n=s.size();
+        int b=0;
         for(int i=n-1; i>=0; i--){
-            long long j=s[i]-'0';
+            int j=s[i]-'0';
             ans += j*pow(2,b);
             b++;
         }
 
-        return (int)ans;
+        return ans;
 
     }
 };
